refactor(CrossAppLogin): Update ExternalAccount prior to 1st public release
Use a Long for the id, to not have to break compatibility if it ever grows.
Reorder properties, and change the proto numbers to be ordered while we can.

+++ CrossAppLoginUI/src/main/kotlin/com/infomaniak/core/crossloginui/views/CrossLoginSelectAccountsView.kt
@@ -46,7 +46,7 @@ class CrossLoginSelectAccountsView @JvmOverloads constructor(
 ) : AbstractComposeView(context, attrs, defStyleAttr) {
 
     private val accounts = mutableStateListOf<ExternalAccount>()
-    private val skippedIds = mutableStateSetOf<Int>()
+    private val skippedIds = mutableStateSetOf<Long>()
 
     private var onClickListener: OnClickListener? = null
 
@@ -113,7 +113,7 @@ class CrossLoginSelectAccountsView @JvmOverloads constructor(
         }
     }
 
-    fun setSkippedIds(items: Set<Int>) {
+    fun setSkippedIds(items: Set<Long>) {
         skippedIds.apply {
             clear()
             addAll(items)
